Add xfce4 keyboard shortcuts file
This change fixes https://github.com/Zegram2112/dotfiles/issues/3, <Alt-F1> was linked to xfce4-popup-applicationsmenu, which is not in ArchLabs. The fix consisted in changing the value for <Alt>F1 to "rofi_run -r" .

This also gives super the current functionality of super+d. This may allow changing the function of super+d.